added leaderboard table to weeklyscores
